// GenericCmsClient/src/app/api/model/entityDescriptor.ts
/**
 * GenericCms
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DynamicProperty } from './dynamicProperty';
import { OperationDescriptor } from './operationDescriptor';


export interface EntityDescriptor { 
    model?: Array<DynamicProperty> | null;
    operations?: Array<OperationDescriptor> | null;
}