// GenericCmsClient/src/app/api/model/dynamicProperty.ts
/**
 * GenericCms
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DynamicProperty { 
    name?: string | null;
    validators?: Array<string> | null;
    readonly type?: string | null;
    defaultValue?: any | null;
}